--- dev/sg/internal/run/check.go
@@ -1,7 +1,0 @@
-package run
-
-type Check struct {
-	Name        string `yaml:"-"`
-	Cmd         string `yaml:"cmd"`
-	FailMessage string `yaml:"failMessage"`
-}
