+++ dev/sg/sg.config.example.yaml
@@ -91,15 +91,6 @@ commands:
       NODE_ENV: development
       NODE_OPTIONS: "--max_old_space_size=8192"
 
-checks:
-  docker:
-    cmd: docker version
-    failMessage: "Failed to run 'docker version'. Please make sure Docker is running."
-
-  redis:
-    cmd: redis-cli -p 6379 PING
-    failMessage: 'Failed to connect to Redis on port 6379. Please make sure Redis is running.'
-
 commandsets:
   minimal:
     - frontend
