+++ sg.config.yaml
@@ -700,31 +700,6 @@ commands:
     cmd: yarn --cwd client/browser dev
     install: yarn
 
-checks:
-  docker:
-    checkFunc: docker
-    failMessage: "Failed to run 'docker version'. Please make sure Docker is running."
-
-  docker-compose:
-    cmd: docker-compose version
-    failMessage: "Failed to run 'docker-compose version'. Please make sure Docker is running."
-
-  go:
-    checkFunc: go
-    failMessage: 'Failed to get correct Go version. Run "sg setup" to make sure you have the right version.'
-
-  redis:
-    checkFunc: redis
-    failMessage: 'Failed to connect to Redis on port 6379. Please make sure Redis is running.'
-
-  postgres:
-    checkFunc: postgres
-    failMessage: 'Failed to connect to Postgres database. Make sure environment variables are setup correctly so that psql can connect.'
-
-  git-version:
-    checkFunc: git
-    failMessage: 'Failed to get correct git version. Run "sg setup" to make sure you have the right version.'
-
 defaultCommandset: enterprise
 commandsets:
   oss:
